환전가능목록 조회 시 데이터 타입 변경 int -> long

+++ src/main/java/com/dalbit/money/vo/Mon_EnableOutputVo.java
@@ -23,8 +23,8 @@ public class Mon_EnableOutputVo {
     private int specialCnt;
 
     private int enableCnt;
-    private int totalGold;
+    private long totalGold;
     private int totalSpecialCnt;
-    private int totalExchangeAmt;
+    private long totalExchangeAmt;
 
 }
